move kernel-boot gate to the real boot; raise default limit
Gate the actual kernel boot inside the lazy IPython provisioner (start +
restore + bootstrap) instead of subagent runtime creation, so session
creation is never delayed -- only the first IPython boot waits for a slot.
The boot weight is the cold Python interpreter + import graph, not the
ports; default limit is now min(16, 2*cores), override via
PRIME_AGENT_MAX_CONCURRENT_KERNEL_BOOTS.
